+++ Frontend/src/router/index.ts
@@ -1,20 +1,31 @@
 import { createRouter, createWebHistory } from 'vue-router'
 import type { RouteRecordRaw } from 'vue-router'
-import Test from '@/views/developerlist.vue'
+import DeveloperList from '@/views/developerlist.vue'
+import Dashboard from '@/views/Dashboard.vue'
 import Listings from '@/views/Listings.vue'
+import Accounting from '@/views/Accounting.vue'
 
 const routes: Array<RouteRecordRaw> = [
   {
     path: '/',
     name: 'Landing',
-    component: Test,
+    component: DeveloperList,
+  },
+  {
+    path: '/dashboard',
+    name: 'Dashboard',
+    component: Dashboard
   },
   {
     path: '/listings',
     name: 'Listing',
     component: Listings
-  }
-  
+  },
+  {
+    path: '/accounting',
+    name: 'Accounting',
+    component: Accounting
+  },
 ]
 
 const router = createRouter({