+++ changelog.md
@@ -19,3 +19,21 @@ Just write whatever might be helpful to know
 ### Need to do:
 - mobile-compatible version
 - smallerizable (like destructable idk the word basta the youtube sidebar effect on pc when you click the 3 lines on the top left)
+
+## 12/30/2025 - Angelo (Frontend)
+### Changes:
+- finally made the listings header
+- made it into tailwind
+- router routes properly
+
+### Notes:
+- screw dropdowns, they need to be manually made to design it like sa figma
+- installed [@headlessui for vue](https://headlessui.com/v1/vue/listbox) to use the [Listbox stuff](https://github.com/Wh4t-20/maurealty-agent-portal/blob/6d62cabc9eafad387af8ea3bf5a2634b8b79d9e6/Frontend/src/views/Listings.vue#L167)
+- I need a backend dude to help me pls (for the filter and data)
+
+
+### Need to do:
+- active background for sidebar
+- design the dropdowns
+- change the layout of the listings
+- add the full detail version of the house when clicked